+++ src/client.rs
@@ -39,6 +39,7 @@ pub struct ClientPlugin;
 impl Plugin for ClientPlugin {
     fn build(&self, app: &mut App) {
         app.init_resource::<RepliconClient>()
+            .init_resource::<NetworkStats>()
             .init_resource::<ServerEntityMap>()
             .init_resource::<ServerUpdateTick>()
             .init_resource::<BufferedMutations>()
@@ -171,21 +172,23 @@ pub(super) fn receive_replication(
 
 fn reset(
     mut client: ResMut<RepliconClient>,
+    mut stats: ResMut<NetworkStats>,
     mut update_tick: ResMut<ServerUpdateTick>,
     mut entity_map: ResMut<ServerEntityMap>,
     mut buffered_mutations: ResMut<BufferedMutations>,
     mutate_ticks: Option<ResMut<ServerMutateTicks>>,
-    stats: Option<ResMut<ClientReplicationStats>>,
+    replication_stats: Option<ResMut<ClientReplicationStats>>,
 ) {
     client.clear();
+    *stats = Default::default();
     *update_tick = Default::default();
     entity_map.clear();
     buffered_mutations.clear();
     if let Some(mut mutate_ticks) = mutate_ticks {
         mutate_ticks.clear();
     }
-    if let Some(mut stats) = stats {
-        *stats = Default::default();
+    if let Some(mut replication_stats) = replication_stats {
+        *replication_stats = Default::default();
     }
 }